Started to work on CozoDB integration with Rust.
Unfortunately, I think it will be more complicated
than expected to deal with Rust and Erlang. Rustler
is mainly compatible with Elixir and the first
error I have is related to... decimal dependency, required
by Jason, but the syntax used is not compatible with
rebar3 or rebar3_hex.

Then, I got two solutions:

  1. improve rust-lib-c implementation and use it with
     a classical C NIF. In fact, it could be interesting
     to learn how rust is dealing with external C code.

  2. switch to elixir and use rustler. This method will
     be easier to maintain and I will have something
     working quick.

  3. A bonus one is to fork rustler and make it compatible
     with Erlang (or offering a support for Erlang). Not
     a good choice.

Another problem I will need to solve. The create to
generate Bytecode are not exposed, in particular
the one to parse cozoscript queries. Temporarily
forking cozodb will be needed.